St blacksmith gui (#53)
* kan trykke på bagrunds billedet

* handlers

* test

* treue

* t

* 12.

* økser sat i blacksmith

* ting

* Buy axes

* axe prixe

* blackstore 21

* blacksmith don

* Revert "Merge branch 'master' into ST_blacksmithGUI"

This reverts commit 87a83caf9d1acfd20209c3bf2601263a100b8bf9, reversing
changes made to 4610ec96023549c9d8182e507297752625edf061.

* Revert "Revert "Merge branch 'master' into ST_blacksmithGUI""

This reverts commit 7d558351ba11350318d7b5d03e959215fa700bd6.

* Revert "Revert "Revert "Merge branch 'master' into ST_blacksmithGUI"""

This reverts commit feb92d622b73f7e7bbff25e0a80cd6d2155b3ebb.

* format huller

* changes

+++ WorldOfZuul/src/game_elements/Tree.java
@@ -37,7 +37,7 @@ public abstract class Tree {
      *
      * @param reduceAmount how much damage the tree has taken
      */
-    public void reduceTreeHealth(double reduceAmount) {
+    public void reduceTreeHealth(int reduceAmount) {
         this.treeHealth -= reduceAmount;
     }
